requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncer==0.0.8
autogen==0.8.5
autogen-agentchat==0.5.1
autogen-core==0.5.1
autogen-ext==0.5.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dotenv==0.9.9
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
gdown==5.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipdb==0.13.13
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jsonref==1.1.0
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
numpy==2.2.4
openai==1.70.0
opentelemetry-api==1.31.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyautogen==0.8.5
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
termcolor==3.0.1
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
